dataproject-django-rest-framework/requirements.txt
Django==2.2
django-crispy-forms==1.12.0
django-environ==0.4.5
djangorestframework==3.12.4
psycopg2==2.9.1
psycopg2-binary==2.9.1
pytz==2021.1
sqlparse==0.4.1
